# openapi_manifest.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Rent a car API
  description: API for renting cars
servers:
  - url: https://api.rent-a-car.local

# For all APIs consumer is authenticated via API key. The only exception are API methods for managing API keys for which Basic Auth is used.
security:
  - ApiKeyAuth: []
paths:
...
  /tenants:
    get:
      description: Returns your account's data
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        '401':
          description: Unauthenticated access
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
             $ref: "#/components/schemas/NewTenant"
      responses:
        '201':
          description: New tenant account has been created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        '401':
          description: Unauthenticated access
        '409':
          description: Account with this email address already exists
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  schemas:
...
    CommonPropsTenant:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        zipCode:
          type: string
        city:
          type: string
        country:
          type: string
          enum: [DE, PL, CZ, SK]
        age:
          type: integer
        email:
          type: string
          format: email
      required:
        - name
        - surname
        - zipCode
        - city
        - country
        - email
    NewTenant:
      allOf:
        - $ref: '#/components/schemas/CommonPropsTenant'
        - type: object
          properties:
            username:
              type: string
              format: password
              description: Used as part of Basic Auth
            password:
              type: string
              format: password
              description: Used as part of Basic Auth
          required:
            - username
            - password
    Tenant:
      allOf:
        - $ref: '#/components/schemas/CommonPropsTenant'
        - type: object
          properties:
            id:
              type: string
          required:
            - id
...